Refactor load save functionality in TaskList to Storage object
